requirements.txt
streamlit==1.47.1
sentence-transformers==4.1.0
scikit-learn==1.7.0
pandas==2.3.0
numpy==1.25.2
torch==2.2.2
groq==0.29.0
python-dotenv==1.1.1
